Align hospital consultant package with multi tool layout
