+++ adk/__init__.py
@@ -1,0 +1,5 @@
+"""Initialize the ADK hospital consultant demo."""
+
+from .hospital_consultant.agent import root_agent
+
+__all__ = ["root_agent"]
