+++ adk/hospital_consultant/__init__.py
@@ -1,5 +1,4 @@
 """Hospital consultant multi-agent package."""
-from .agents.team_lead import run_team_lead
-from .context import context_store
 
-__all__ = ["run_team_lead", "context_store"]
+from . import agent
+
